feat(web,wasm): Phase 7/8.4 — interactive multi-scale drill-in navigation
Wire the refinement framework through to the browser: click the map to zoom into
a finer sector, breadcrumb back out.

- mapgen-wasm: `refineSector(seed, plates, level, sx, sy, cells) -> WorldHandle`
  recomputes the shared base field from the seed and refines that quadtree
  sector; the existing `render()` honours its viewBox, so the frontend just swaps
  the SVG. A wasm-bindgen-test renders a sector under wasm (its own
  `512 320 512 320` viewBox), so the browser path is covered, not just native.
- worker: keeps the root world + seed, adds a `refine` message that (re-)refines
  a sector statelessly (navigating up just re-refines the parent — no caching);
  narration still targets the whole world.
- main: a `Nav {level, sx, sy}` quadtree cursor. A near-stationary click maps
  screen → world coordinates and drills into the child sector beneath it; a drag
  still pans. Coarse-first UX (panzoom.focusContentRect zooms into the clicked
  region using the parent's pixels while the sector generates, then swaps),
  clickable breadcrumb (World › L1 › L2 …) for up-navigation, zoom-in cursor.
- panzoom: `clientToContent` (screen → SVG/world coords) + `focusContentRect`
  (animated coarse-first framing).

This realizes ADR 0001's discrete drill-in + overview/detail model. Frontend
`tsc --noEmit` + vite build clean; just check + the wasm golden green.

+++ crates/mapgen-cli/tests/refine_render.rs
@@ -62,7 +62,7 @@ fn refined_sector_renders_its_own_viewport_and_is_not_blank() {
         }
         sum += (r + g + b) / 3.0;
         // Bluish pixels = sea/river.
-        if (b as f64) > r + 12.0 && (b as f64) > g + 6.0 {
+        if b > r + 12.0 && b > g + 6.0 {
             water += 1;
         }
     }
